Added Project -> EEPROM_Event_Counter
Basically a counter that counts and it's state is saved in an address of the EEPROM. It can also be reset so you can start counting again

This Arduino sketch implements a reliable event counter. Its main function is to use a single push button to increment a count, while using the Arduino's internal EEPROM (Electrically Erasable Programmable Read-Only Memory) to save the total count. This ensures the data persists even when the power is disconnected.